RISC-V: rename parse_asm.h to insn.h
commit bb5fc504e59489c5b32a893f72791da6e992e6b0 from
https://github.com/linux4microchip/linux.git linux-6.1-mchp+fpga

The current parse_asm header should become a more centralized place
for everything concerning parsing and constructing instructions.

We already have a header insn-def.h similar to aarch64, so rename
parse_asm.h to insn.h (again similar to aarch64) to show that it's
meant for more than simple instruction parsing.

Link: https://lore.kernel.org/r/20221223221332.4127602-8-heiko@sntech.de
